:alien: sample-admin 适配 ballcat v1.3.0 版本的改动

+++ ballcat-sample-admin-application/sql/sample.sql
@@ -1,9 +1,9 @@
 -- 文档表，用于演示数据权限
 CREATE TABLE `sample_document` (
-                                   `id` int(255) NOT NULL AUTO_INCREMENT COMMENT 'ID',
+                                   `id` bigint(20) NOT NULL AUTO_INCREMENT COMMENT 'ID',
                                    `name` varchar(50) DEFAULT NULL COMMENT '文档名称',
-                                   `user_id` int(11) DEFAULT NULL COMMENT '所属用户ID',
-                                   `organization_id` int(11) DEFAULT NULL COMMENT '所属组织ID',
+                                   `user_id` bigint(20) DEFAULT NULL COMMENT '所属用户ID',
+                                   `organization_id`  bigint(20) DEFAULT NULL COMMENT '所属组织ID',
                                    `create_time` datetime DEFAULT NULL COMMENT '创建时间',
                                    `update_time` datetime DEFAULT NULL COMMENT '更新时间',
                                    PRIMARY KEY (`id`)
